refactor: separate fitting from prtecan and define __all__
