+++ pyproject.toml
@@ -119,7 +119,7 @@ sync = "pip-df sync -x tests"  # XXX:
 test = "coverage run --parallel -m pytest"
 type = "mypy src tests docs/conf.py"
 xdoc = "python -m xdoctest clophfit all"
-all = ["sync", "test", "type", "xdoc"]
+all = ["test", "type", "xdoc"]
 [[tool.hatch.envs.tests.matrix]]
 python = ["3.11", "3.10", "3.9", "3.8"] #reverse order to ensure the presence in older python of module included in newer versions
 
@@ -323,14 +323,6 @@ messages_control.disable = [
 # messages_control.enable = ["raise-missing-from"]
 
 [tool.ruff]
-ignore = [
-    "ANN001",
-    "ANN101",
-    "ANN102",
-    "ANN201",
-    "E501",
-    "S101",
-]
 line-length = 88
 select = [
     "A",                        # builtins
@@ -355,6 +347,7 @@ select = [
     "N",                        # pep8-naming
     "PD",                       # pandas-vet
     # "PGH",  # pygrep WAIT
+    # "PL",                       # pylint see PLR2004...
     "PT",                       # pytest-style
     "PTH",                      # use-pathlib
     "Q",                        # quotes
@@ -368,6 +361,12 @@ select = [
 ]
 extend-exclude = ["src/clophfit/old/"]
 force-exclude = true
+ignore = [
+    "ANN101",
+    "ANN102",
+    "ANN201",
+    "E501",
+]
 
 [tool.ruff.mccabe]
 max-complexity = 33
@@ -381,15 +380,13 @@ force-single-line = true
 # single-line-exclusions = ["os", "json"]
 split-on-trailing-comma = true
 known-first-party = ["src", "clophfit"]
-# required-imports = ["from __future__ import annotations"]
+required-imports = ["from __future__ import annotations"]
 
 [tool.ruff.per-file-ignores]
 "tests/*" = [
-    "ANN101",  # XXX:
-    "ANN102",
-    "S101",
+    "PLR2004",
 ]
-"src/clophfit/prtecan.py" = [
-    "B905",
-    "E741",
+"src/clophfit/__main__.py" = [
+    "ANN001",                   # params are annotated using click
 ]
+"__init__.py" = ["I002"]        # don't need annotations
