+++ src/clophfit/binding/fitting.py
@@ -1,0 +1,209 @@
+"""Fit Cl binding and pH titration."""
+from __future__ import annotations
+
+from typing import Sequence
+
+import numpy as np
+import pandas as pd
+import scipy  # type: ignore
+import scipy.stats  # type: ignore
+from numpy.typing import NDArray
+
+"""Binding equations."""
+# TODO: use this like fz in prtecan
+
+
+def kd(
+    kd1: float, pka: float, ph: NDArray[np.float_] | float
+) -> NDArray[np.float_] | float:
+    """Infinite cooperativity model.
+
+    It can describe pH-dependence for chloride dissociation constant.
+
+    Parameters
+    ----------
+    kd1 : float
+        Dissociation constant at pH <= 5.0 (fully protonated).
+    pka : float
+        Acid dissociation constant.
+    ph : Xtype
+        pH value(s).
+
+    Returns
+    -------
+    Xtype
+        Predicted Kd value(s).
+
+    Examples
+    --------
+    >>> kd(10, 8.4, 7.4)
+    11.0
+    >>> import numpy as np
+    >>> kd(10, 8.4, np.array([7.4, 8.4]))
+    array([11., 20.])
+
+    """
+    return kd1 * (1 + 10 ** (pka - ph)) / 10 ** (pka - ph)
+
+
+# TODO other from datan
+
+
+def fz_kd_singlesite(
+    k: float, p: NDArray[np.float_] | Sequence[float], x: NDArray[np.float_]
+) -> NDArray[np.float_]:
+    """Fit function for Cl titration."""
+    return (float(p[0]) + float(p[1]) * x / k) / (1 + x / k)
+
+
+def fz_pk_singlesite(
+    k: float, p: NDArray[np.float_] | Sequence[float], x: NDArray[np.float_]
+) -> NDArray[np.float_]:
+    """Fit function for pH titration."""
+    return (float(p[1]) + float(p[0]) * 10 ** (k - x)) / (1 + 10 ** (k - x))
+
+
+def fit_titration(
+    kind: str,
+    x: Sequence[float],
+    y: NDArray[np.float_],
+    y2: NDArray[np.float_] | None = None,
+    residue: NDArray[np.float_] | None = None,
+    residue2: NDArray[np.float_] | None = None,
+    tval_conf: float = 0.95,
+) -> pd.DataFrame:
+    """Fit pH or Cl titration using a single-site binding model.
+
+    Returns confidence interval (default=0.95) for fitting params (cov*tval), rather than
+    standard error of the fit. Use scipy leastsq. Determine 3 fitting parameters:
+    - binding constant *K*
+    - and 2 plateau *SA* and *SB*.
+
+    Parameters
+    ----------
+    kind : str
+        Titration type {'pH'|'Cl'}
+    x : Sequence[float]
+        Dataset x-values.
+    y : NDArray[np.float]
+        Dataset y-values.
+    y2 : NDArray[np.float], optional
+        Optional second dataset y-values (share x with main dataset).
+    residue : NDArray[np.float], optional
+        Residues for main dataset.
+    residue2 : NDArray[np.float], optional
+        Residues for second dataset.
+    tval_conf : float
+        Confidence level (default 0.95) for parameter estimations.
+
+    Returns
+    -------
+    pd.DataFrame
+        Fitting results.
+
+    Raises
+    ------
+    NameError
+        When kind is different than "pH" or "Cl".
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> fit_titration("Cl", np.array([1.0, 10, 30, 100, 200]), \
+          np.array([10, 8, 5, 1, 0.1]))[["K", "sK"]]
+               K         sK
+    0  38.955406  30.201929
+
+    """
+    if kind == "pH":
+        fz = fz_pk_singlesite
+    elif kind == "Cl":
+        fz = fz_kd_singlesite
+    else:
+        msg = "kind= pH or Cl"
+        raise NameError(msg)
+
+    def compute_p0(x: Sequence[float], y: NDArray[np.float_]) -> NDArray[np.float_]:
+        data = pd.DataFrame({"x": x, "y": y})
+        p0sa = data.y[data.x == min(data.x)].to_numpy()[0]
+        p0sb = data.y[data.x == max(data.x)].to_numpy()[0]
+        p0k = np.average([max(y), min(y)])
+        try:
+            x1, y1 = data[data["y"] >= p0k].to_numpy()[0]
+        except IndexError:
+            x1 = np.nan
+            y1 = np.nan
+        try:
+            x2, y2 = data[data["y"] <= p0k].to_numpy()[0]
+        except IndexError:
+            x2 = np.nan
+            y2 = np.nan
+        p0k = (x2 - x1) / (y2 - y1) * (p0k - y1) + x1
+        return np.array(np.r_[p0k, p0sa, p0sb])
+
+    if y2 is None:
+
+        def ssq1(
+            p: NDArray[np.float_], x: NDArray[np.float_], y1: NDArray[np.float_]
+        ) -> NDArray[np.float_]:
+            return np.array(np.r_[y1 - fz(p[0], p[1:3], x)])
+
+        p0 = compute_p0(x, y)
+        p, cov, info, msg, success = scipy.optimize.leastsq(
+            ssq1, p0, args=(np.array(x), y), full_output=True, xtol=1e-11
+        )
+    else:
+
+        def ssq2(
+            p: NDArray[np.float_],
+            x: NDArray[np.float_],
+            y1: NDArray[np.float_],
+            y2: NDArray[np.float_],
+            rd1: NDArray[np.float_],
+            rd2: NDArray[np.float_],
+        ) -> NDArray[np.float_]:
+            return np.array(
+                np.r_[
+                    (y1 - fz(p[0], p[1:3], x)) / rd1**2,
+                    (y2 - fz(p[0], p[3:5], x)) / rd2**2,
+                ]
+            )
+
+        p1 = compute_p0(x, y)
+        p2 = compute_p0(x, y2)
+        ave = np.average([p1[0], p2[0]])
+        p0 = np.r_[ave, p1[1], p1[2], p2[1], p2[2]]
+        tmp = scipy.optimize.leastsq(
+            ssq2,
+            p0,
+            full_output=True,
+            xtol=1e-11,
+            args=(np.array(x), y, y2, residue, residue2),
+        )
+        p, cov, info, msg, success = tmp
+    res = pd.DataFrame({"ss": [success]})
+    res["msg"] = msg
+    if 1 <= success <= 4:
+        try:
+            tval = (tval_conf + 1) / 2
+            chisq = sum(info["fvec"] * info["fvec"])
+            res["df"] = len(y) - len(p)
+            res["tval"] = scipy.stats.distributions.t.ppf(tval, res.df)
+            res["chisqr"] = chisq / res.df
+            res["K"] = p[0]
+            res["SA"] = p[1]
+            res["SB"] = p[2]
+            if y2 is not None:
+                res["df"] += len(y2)
+                res["tval"] = scipy.stats.distributions.t.ppf(tval, res.df)
+                res["chisqr"] = chisq / res.df
+                res["SA2"] = p[3]
+                res["SB2"] = p[4]
+                res["sSA2"] = np.sqrt(cov[3][3] * res.chisqr) * res.tval
+                res["sSB2"] = np.sqrt(cov[4][4] * res.chisqr) * res.tval
+            res["sK"] = np.sqrt(cov[0][0] * res.chisqr) * res.tval
+            res["sSA"] = np.sqrt(cov[1][1] * res.chisqr) * res.tval
+            res["sSB"] = np.sqrt(cov[2][2] * res.chisqr) * res.tval
+        except TypeError:
+            pass  # if some params are not successfully determined.
+    return res
